Modified the file 100-times_table.c and used a different approach with two nested for loops and two signed long integers values i and j and prints the result in format of %d

+++ 0x02-functions_nested_loops/100-times_table.c
@@ -13,7 +13,7 @@
  */
 void print_times_table(int n)
 {
-	signed long int i, j;
+	int i, j;
 
 	if (n < 0 || n > 15)
 	{
@@ -24,7 +24,7 @@ void print_times_table(int n)
 	{
 		for (j = 0; j <= 12; j++)
 		{
-			printf("%-4d,  ", i * j);
+			printf("%d,  ", i * j);
 		}
 		printf("\n");
 	}
